Add files via upload
Hi Everyone,
Welcome to my facial recognition project repository! I'm excited to share my journey in the world of computer vision and machine learning with you. With 1 year of hands-on experience in this field, I've crafted this project to showcase the power and potential of facial recognition technology. Throughout my career, I've honed my skills in Python, OpenCV, TensorFlow, and DLib, among other technologies, to develop robust solutions for real-world challenges. Whether you're a beginner eager to learn or an experienced developer looking for inspiration, I hope this repository serves as a valuable resource for you. Feel free to explore, contribute, and embark on your own journey into the fascinating realm of facial recognition!