Add files via upload
Hi Everyone,

Welcome to my PlayStation clone website repository! Delve into the realm of gaming nostalgia with this project inspired by the iconic PlayStation console. With 6 months of experience in web development, I've meticulously crafted this site to capture the essence of the beloved gaming platform. From sleek design elements to intuitive user interfaces, every aspect is designed to evoke the excitement and joy of gaming. Whether you're a seasoned gamer or a curious enthusiast, I invite you to explore this repository and relive the magic of PlayStation. Join me on this journey into the world of gaming excellence, and let's celebrate the legacy of PlayStation together!